move minitests to test/ directory
