+++ test-viz.rb
@@ -35,9 +35,9 @@ end
 require 'minitest/autorun'
 require 'minitest/reporters'
 Minitest::Reporters.use! [Minitest::Reporters::ProgressReporter.new(:color => true)]
-require './test-loader'
-require './test-postprocessor'
-require './test-resolver'
-require './test-varfinder'
-require './test-scoper'
-require './test-grapher'
+require './test/loader'
+require './test/postprocessor'
+require './test/resolver'
+require './test/varfinder'
+require './test/scoper'
+require './test/grapher'
