+++ test/grapher.rb
