+++ test/loader.rb
