+++ test/resolver.rb
