+++ test/scoper.rb
