+++ test/varfinder.rb
